Split out test-helper
